Tabs are replaced with spaces, failed notifications store renamed
