+++ src/Journalist.EventStore/Configuration/EventStoreConnectionConfiguration.cs
@@ -54,21 +54,21 @@ namespace Journalist.EventStore.Configuration
         }
 
         public IEventStoreConnectionConfiguration UseStorage(
-			string storageConnectionString, 
-			string journalTableName = Constants.StorageEntities.EVENT_JOURNAL_TABLE_NAME, 
-			string eventStoreDeploymentTableName = Constants.StorageEntities.EVENT_STORE_DEPLOYMENT_TABLE_NAME, 
-			string failedNotificationsTableName = Constants.StorageEntities.FAILED_NOTIFICATIONS_TABLE_NAME,
-			string notificationQueueName = Constants.StorageEntities.NOTIFICATION_QUEUE_NAME, 
-			int notificationQueuePartitionCount = Constants.Settings.NOTIFICATION_QUEUE_PARTITION_COUNT,
-			string streamConsumerSessionsBlobContainerName = Constants.StorageEntities.EVENT_CONSUMER_SESSIONS_BLOB_CONTAINER_NAME,
-			string pendingNotificationsTableName = Constants.StorageEntities.PENDING_NOTIFICATIONS_TABLE_NAME, 
-			string pendingNotificationsChaserExclusiveAccessLockBlobContainerName = Constants.StorageEntities.PENDING_NOTIFICATIONS_CHASER_EXCLUSIVE_ACCESS_LOCK_BLOB_CONTAINER_NAME, 
-			string pendingNotificationsChaserExclusiveAccessLockBlobName = Constants.StorageEntities.PENDING_NOTIFICATIONS_CHASER_EXCLUSIVE_ACCESS_LOCK_BLOB_NAME)
+            string storageConnectionString, 
+            string journalTableName = Constants.StorageEntities.EVENT_JOURNAL_TABLE_NAME, 
+            string eventStoreDeploymentTableName = Constants.StorageEntities.EVENT_STORE_DEPLOYMENT_TABLE_NAME, 
+            string failedNotificationsTableName = Constants.StorageEntities.FAILED_NOTIFICATIONS_TABLE_NAME,
+            string notificationQueueName = Constants.StorageEntities.NOTIFICATION_QUEUE_NAME, 
+            int notificationQueuePartitionCount = Constants.Settings.NOTIFICATION_QUEUE_PARTITION_COUNT,
+            string streamConsumerSessionsBlobContainerName = Constants.StorageEntities.EVENT_CONSUMER_SESSIONS_BLOB_CONTAINER_NAME,
+            string pendingNotificationsTableName = Constants.StorageEntities.PENDING_NOTIFICATIONS_TABLE_NAME, 
+            string pendingNotificationsChaserExclusiveAccessLockBlobContainerName = Constants.StorageEntities.PENDING_NOTIFICATIONS_CHASER_EXCLUSIVE_ACCESS_LOCK_BLOB_CONTAINER_NAME, 
+            string pendingNotificationsChaserExclusiveAccessLockBlobName = Constants.StorageEntities.PENDING_NOTIFICATIONS_CHASER_EXCLUSIVE_ACCESS_LOCK_BLOB_NAME)
         {
             Require.NotEmpty(storageConnectionString, "storageConnectionString");
             Require.NotEmpty(journalTableName, "journalTableName");
             Require.NotEmpty(eventStoreDeploymentTableName, "eventStoreDeploymentTableName");
-			Require.NotEmpty(failedNotificationsTableName, nameof(failedNotificationsTableName));
+            Require.NotEmpty(failedNotificationsTableName, nameof(failedNotificationsTableName));
             Require.NotEmpty(notificationQueueName, "notificationQueueName");
             Require.Positive(notificationQueuePartitionCount, "notificationQueuePartitionCount");
             Require.NotEmpty(streamConsumerSessionsBlobContainerName, "streamConsumerSessionsBlobContainerName");
@@ -79,7 +79,7 @@ namespace Journalist.EventStore.Configuration
             StorageConnectionString = storageConnectionString;
             JournalTableName = journalTableName;
             EventStoreDeploymentTableName = eventStoreDeploymentTableName;
-	        FailedNotificationsTableName = failedNotificationsTableName;
+            FailedNotificationsTableName = failedNotificationsTableName;
             NotificationQueueName = notificationQueueName;
             NotificationQueuePartitionCount = notificationQueuePartitionCount;
             StreamConsumerSessionsBlobContainerName = streamConsumerSessionsBlobContainerName;
@@ -89,8 +89,8 @@ namespace Journalist.EventStore.Configuration
 
             return this;
         }
-		
-	    public IEventMutationPipelineConfiguration Mutate
+        
+        public IEventMutationPipelineConfiguration Mutate
         {
             get { return m_mutationPipelineConfiguration; }
         }
@@ -116,15 +116,15 @@ namespace Journalist.EventStore.Configuration
         {
             get;
             private set;
-		}
+        }
 
-	    public string FailedNotificationsTableName
-	    {
-		    get;
-			private set;
-	    }
+        public string FailedNotificationsTableName
+        {
+            get;
+            private set;
+        }
 
-		public string NotificationQueueName
+        public string NotificationQueueName
         {
             get;
             private set;
