+++ src/Journalist.EventStore/Notifications/Listeners/NotificationListenerSubscription.cs
@@ -17,23 +17,23 @@ namespace Journalist.EventStore.Notifications.Listeners
 
         private readonly INotificationsChannel m_notificationsChannel;
         private readonly INotificationListener m_listener;
-	    private readonly IFailedNotificationsHub m_failedNotificationsHub;
-	    private readonly CountdownEvent m_processingCountdown;
+        private readonly IFailedNotificationsStore m_failedNotificationsStore;
+        private readonly CountdownEvent m_processingCountdown;
         private IEventStoreConnection m_connection;
 
         public NotificationListenerSubscription(
             INotificationsChannel notificationsChannel,
             INotificationListener listener,
-			IFailedNotificationsHub failedNotificationsHub)
+            IFailedNotificationsStore failedNotificationsStore)
         {
             Require.NotNull(notificationsChannel, nameof(notificationsChannel));
             Require.NotNull(listener, nameof(listener));
-			Require.NotNull(failedNotificationsHub, nameof(failedNotificationsHub));
+            Require.NotNull(failedNotificationsStore, nameof(failedNotificationsStore));
 
             m_notificationsChannel = notificationsChannel;
             m_listener = listener;
-	        m_failedNotificationsHub = failedNotificationsHub;
-	        m_processingCountdown = new CountdownEvent(0);
+            m_failedNotificationsStore = failedNotificationsStore;
+            m_processingCountdown = new CountdownEvent(0);
         }
 
         public async Task HandleNotificationAsync(INotification notification)
@@ -125,7 +125,7 @@ namespace Journalist.EventStore.Notifications.Listeners
                     notification.DeliveryCount.ToInvariantString(),
                     Constants.Settings.MAX_NOTIFICATION_PROCESSING_ATTEMPT_COUNT.ToInvariantString());
 
-	            await m_failedNotificationsHub.PutToFailedAsync(notification);
+                await m_failedNotificationsStore.PutToFailedAsync(notification);
             }
         }
