+++ src/Journalist.EventStore/Notifications/NotificationHub.cs
@@ -21,25 +21,25 @@ namespace Journalist.EventStore.Notifications
         private readonly IPollingJob m_pollingJob;
         private readonly INotificationsChannel m_channel;
         private readonly IReceivedNotificationProcessor m_notificationProcessor;
-	    private readonly IFailedNotificationsHub m_failedNotificationsHub;
+        private readonly IFailedNotificationsStore m_failedNotificationsStore;
 
-	    private int m_maxProcessingCount;
+        private int m_maxProcessingCount;
 
         public NotificationHub(
             IPollingJob pollingJob,
             INotificationsChannel channel,
             IReceivedNotificationProcessor notificationProcessor,
-			IFailedNotificationsHub failedNotificationsHub)
+            IFailedNotificationsStore failedNotificationsStore)
         {
             Require.NotNull(pollingJob, nameof(pollingJob));
             Require.NotNull(channel, nameof(channel));
             Require.NotNull(notificationProcessor, nameof(notificationProcessor));
-			Require.NotNull(failedNotificationsHub, nameof(failedNotificationsHub));
+            Require.NotNull(failedNotificationsStore, nameof(failedNotificationsStore));
 
             m_pollingJob = pollingJob;
             m_channel = channel;
             m_notificationProcessor = notificationProcessor;
-	        m_failedNotificationsHub = failedNotificationsHub;
+            m_failedNotificationsStore = failedNotificationsStore;
         }
 
         public Task NotifyAsync(INotification notification)
@@ -53,8 +53,8 @@ namespace Journalist.EventStore.Notifications
         {
             Require.NotNull(listener, "listener");
 
-			m_failedNotificationsHub.AddRetryListener(listener);
-            m_subscriptions.Add(listener.GetType(), new NotificationListenerSubscription(m_channel, listener, m_failedNotificationsHub));
+            m_failedNotificationsStore.AddRetryListener(listener);
+            m_subscriptions.Add(listener.GetType(), new NotificationListenerSubscription(m_channel, listener, m_failedNotificationsStore));
         }
 
         public void Unsubscribe(INotificationListener listener)
@@ -85,7 +85,7 @@ namespace Journalist.EventStore.Notifications
 
                     if (notifications.IsEmpty())
                     {
-                        s_logger.Debug("No notifications for processing.");
+                        s_logger.Verbose("No notifications for processing.");
 
                         return false;
                     }
