+++ src/Journalist.EventStore/Streams/EventStreamConsumer.cs
@@ -69,16 +69,16 @@ namespace Journalist.EventStore.Streams
                 await m_commitConsumedVersion(version);
                 m_stateMachine.ConsumedStreamVersionCommited(version, skipCurrent);
             }
-		}
+        }
 
-		public Task CommitStreamVersionAsync(StreamVersion streamVersion)
-		{
-			Require.NotNull(streamVersion, "streamVersion");
+        public Task CommitStreamVersionAsync(StreamVersion streamVersion)
+        {
+            Require.NotNull(streamVersion, "streamVersion");
 
-			return m_commitConsumedVersion(streamVersion);
-		}
+            return m_commitConsumedVersion(streamVersion);
+        }
 
-		public async Task CloseAsync()
+        public async Task CloseAsync()
         {
             if (m_stateMachine == null)
             {
@@ -125,6 +125,6 @@ namespace Journalist.EventStore.Streams
 
         public string StreamName => m_reader.StreamName;
 
-	    public StreamVersion CurrentConsumerVersion => m_stateMachine.CommitedStreamVersion;
+        public StreamVersion CurrentConsumerVersion => m_stateMachine.CommitedStreamVersion;
     }
 }
