+++ src/Journalist.EventStore/Streams/EventStreamConsumers.cs
@@ -54,34 +54,34 @@ namespace Journalist.EventStore.Streams
             m_cache.TryAdd(consumerName, consumerId);
 
             return consumerId;
-		}
+        }
 
-		public async Task<IEnumerable<EventStreamReaderId>> EnumerateAsync()
-		{
-			var query = m_consumerMetadataTable.PrepareEntitySegmentedRangeQueryByPartition(
-				Constants.StorageEntities.MetadataTable.EVENT_STREAM_CONSUMERS_IDS_PK);
+        public async Task<IEnumerable<EventStreamReaderId>> EnumerateAsync()
+        {
+            var query = m_consumerMetadataTable.PrepareEntitySegmentedRangeQueryByPartition(
+                Constants.StorageEntities.MetadataTable.EVENT_STREAM_CONSUMERS_IDS_PK);
 
-			var result = await query.ExecuteAsync();
-			if (result == null)
-			{
-				return Enumerable.Empty<EventStreamReaderId>();
-			}
+            var result = await query.ExecuteAsync();
+            if (result == null)
+            {
+                return Enumerable.Empty<EventStreamReaderId>();
+            }
 
-			var eventStreamReaderIds = new List<EventStreamReaderId>();
-			do
-			{
-				eventStreamReaderIds.AddRange(
-					result
-						.Where(row => row.ContainsKey(Constants.StorageEntities.MetadataTableProperties.EVENT_STREAM_CONSUMER_NAME))
-						.Select(row => EventStreamReaderId.Parse((string)row[KnownProperties.RowKey])));
-				result = await query.ExecuteAsync(query.ContinuationToken);
-			}
-			while (query.HasMore);
+            var eventStreamReaderIds = new List<EventStreamReaderId>();
+            do
+            {
+                eventStreamReaderIds.AddRange(
+                    result
+                        .Where(row => row.ContainsKey(Constants.StorageEntities.MetadataTableProperties.EVENT_STREAM_CONSUMER_NAME))
+                        .Select(row => EventStreamReaderId.Parse((string)row[KnownProperties.RowKey])));
+                result = await query.ExecuteAsync(query.ContinuationToken);
+            }
+            while (query.HasMore);
 
-			return eventStreamReaderIds;
-		}
+            return eventStreamReaderIds;
+        }
 
-		private async Task<EventStreamReaderId> QueryConsumerId(string consumerName)
+        private async Task<EventStreamReaderId> QueryConsumerId(string consumerName)
         {
             var query = m_consumerMetadataTable.PrepareEntityPointQuery(
                 partitionKey: Constants.StorageEntities.MetadataTable.EVENT_STREAM_CONSUMERS_IDS_PK,
